problem 8 toegevoegd en interface methode getanswer omgeschreven naar return van een long zodat grotere getallen worden ondersteund

+++ MathBarld/ProjectEuler/IProblem.cs
@@ -3,7 +3,7 @@ namespace MathBarld.ProjectEuler
 {
     public interface IProblem
     {
-        int GetAnswer();
+        long GetAnswer();
         string ProblemDescription { get; }
     }
 }
