+++ MathBarld/ProjectEuler/Problem5.cs
@@ -12,7 +12,7 @@ namespace MathBarld.ProjectEuler
 
 What is the smallest positive number that is evenly divisible by all of the numbers from 1 to 20?";
 
-        public int GetAnswer()
+        public long GetAnswer()
         {
             for (int i = 20; ; i += 20)
             {
@@ -26,7 +26,7 @@ What is the smallest positive number that is evenly divisible by all of the numb
                     }
                 }
                 if (divisable)
-                    return i;
+                    return (long)i;
             }
         }
     }
